API Reference

API Reference

Types

Functions

Types

Missing docstring.

Missing docstring for AbstractRGBE. Check Documenter's build log for details.

Missing docstring.

Missing docstring for AbstractXYZE. Check Documenter's build log for details.

Missing docstring.

Missing docstring for AbstractCCCE. Check Documenter's build log for details.

Missing docstring.

Missing docstring for AbstractRGBE32. Check Documenter's build log for details.

Missing docstring.

Missing docstring for AbstractXYZE32. Check Documenter's build log for details.

Missing docstring.

Missing docstring for AbstractCCCE32. Check Documenter's build log for details.

HDRColorTypes.RGBE32 — Type

RGBE32{T<:AbstractFloat} <: AbstractRGBE32{T}

32-bit RGBE color type based on the Radiance HDR format. Note that ColorTypes.RGB is usually assumed to be gamma-corrected sRGB, but the RGBE format usually handles linear RGBs.

HDRColorTypes.XYZE32 — Type

XYZE32{T} <: AbstractXYZE32{T}

32-bit XYZE color type based on the Radiance HDR format.

HDRColorTypes.CCCE32 — Type

CCCE32{T} = Union{RGBE32{T}, XYZE32{T}}

Functions

HDRColorTypes.rgbe32 — Function

rgbe32(r8::UInt8, g8::UInt8, b8::UInt8, e8::UInt8) -> RGBE32{Float32}

Construct an RGBE32 color from four arguments in UInt8. Note that RGBE32(r8, g8, b8, e8) is not allowed for compatibility with other AbstractRGB and TransparentRGB color constructors.

See also ccce32.

rgbe32(rgbe::UInt32) -> RGBE32{Float32}

Construct an RGBE32 color from a UInt32 value in 0xRRGGBBEE format.

See also ccce32.

HDRColorTypes.xyze32 — Function

xyze32(x8::UInt8, y8::UInt8, z8::UInt8, e8::UInt8) -> XYZE32{Float32}

Construct an XYZE32 color from four arguments of UInt8. Note that XYZE32(x8, y8, z8, e8) is not allowed for compatibility with other XYZ, AXYZ, and XYZA color constructors.

See also ccce32.

xyze32(xyze::UInt32) -> XYZE32{Float32}

Construct an XYZE32 color from a UInt32 value in 0xXXYYZZEE format.

See also ccce32.

HDRColorTypes.ccce32 — Function

ccce32(RGBE32{T}, c1::UInt8, c2::UInt8, c3::UInt8, e::UInt8) -> RGBE32{T}
ccce32(XYZE32{T}, c1::UInt8, c2::UInt8, c3::UInt8, e::UInt8) -> XYZE32{T}

Construct a Radiance HDR format color from four arguments of UInt8. If the component type T is not specified, Float32 is applied.

This function is a generic version of rgbe32 and xyze32.

ccce32(RGBE32{T}, ccce::UInt32) -> RGBE32{T}
ccce32(XYZE32{T}, ccce::UInt32) -> XYZE32{T}

Construct a Radiance HDR format color from a UInt32 value in 0xRRGGBBEE or 0xXXYYZZEE format. If the component type T is not specified, Float32 is applied.

This function is a generic version of rgbe32 and xyze32.